Conflicts:
	build/cxx_tests.rb
	build/nginx.rb
	ext/common/agents/HelperAgent/HttpStatusExtractor.h
	ext/common/agents/HelperAgent/Main.cpp

+++ ext/common/AccountsDatabase.h
@@ -61,7 +61,7 @@ public:
 	
 	unsigned int size() const {
 		lock_guard<boost::mutex> l(lock);
-		return accounts.size();
+		return (unsigned int) accounts.size();
 	}
 	
 	vector<string> listUsernames() const {
