+++ ext/common/Session.h
@@ -176,7 +176,7 @@ public:
 	 * @throws boost::thread_interrupted
 	 */
 	virtual void sendHeaders(const StaticString &headers) {
-		sendHeaders(headers.c_str(), headers.size());
+		sendHeaders(headers.c_str(), (unsigned int) headers.size());
 	}
 	
 	/**
